update text context pair script
